+++ src/main/java/oshi/software/os/mac/MacHardwareAbstractionLayer.java
@@ -61,8 +61,7 @@ public class MacHardwareAbstractionLayer implements HardwareAbstractionLayer {
 	public float getProcessorLoad() {
 		ArrayList<String> topResult = ExecutingCommand.runNative("top -l 1 -R -F -n1"); // cpu load is in [3]
 		String[] idle = topResult.get(3).split(" "); // idle value is in [6]
-		float load = 100 - Float.valueOf(idle[6].replace("%", ""));
-		return load;
+		return 100 - Float.valueOf(idle[6].replace("%", ""));
 	}
 
 }