+++ src/main/java/oshi/hardware/HardwareAbstractionLayer.java
@@ -26,4 +26,5 @@ public interface HardwareAbstractionLayer {
 	 *  A memory object.
 	 */
 	Memory getMemory();
+
 }