+++ src/main/java/oshi/software/os/mac/local/CentralProcessor.java
@@ -10,6 +10,8 @@ package oshi.software.os.mac.local;
 import oshi.hardware.Processor;
 import oshi.util.ExecutingCommand;
 
+import java.util.ArrayList;
+
 /**
  * A CPU.
  * 
@@ -177,6 +179,15 @@ public class CentralProcessor implements Processor {
 	 */
 	public void setFamily(String _family) {
 		this._family = _family;
+	}
+
+	/**
+	 * {@inheritDoc}
+	 */
+	public float getLoad() {
+		ArrayList<String> topResult = ExecutingCommand.runNative("top -l 1 -R -F -n1"); // cpu load is in [3]
+		String[] idle = topResult.get(3).split(" "); // idle value is in [6]
+		return 100 - Float.valueOf(idle[6].replace("%", ""));
 	}
 
 	@Override