+++ src/main/java/oshi/software/os/windows/WindowsHardwareAbstractionLayer.java
@@ -11,6 +11,7 @@ import oshi.software.os.windows.nt.GlobalMemory;
 
 import com.sun.jna.platform.win32.Advapi32Util;
 import com.sun.jna.platform.win32.WinReg;
+import oshi.util.ExecutingCommand;
 
 public class WindowsHardwareAbstractionLayer implements HardwareAbstractionLayer {