+++ src/test/java/oshi/SystemInfoTest.java
@@ -46,6 +46,13 @@ public class SystemInfoTest {
         assertTrue(memory.getAvailable() <= memory.getTotal());
     }
 
+    @Test
+    public void testCpuLoad() {
+        SystemInfo si = new SystemInfo();
+        HardwareAbstractionLayer hal = si.getHardware();
+        assertTrue(hal.getProcessors()[0].getLoad() >= 0 && hal.getProcessors()[0].getLoad() <= 100);
+    }
+
     public static void main(String[] args) {
         SystemInfo si = new SystemInfo();
         // software
@@ -63,6 +70,7 @@ public class SystemInfoTest {
         System.out.println("Memory: "
                 + FormatUtil.formatBytes(hal.getMemory().getAvailable()) + "/"
                 + FormatUtil.formatBytes(hal.getMemory().getTotal()));
+        System.out.println("CPU load: " + hal.getProcessors()[0].getLoad() + "%");
     }
 
 }